add delete button
also make border invisible instead of grow on the title and editor.  prevents words from moving around when focusing them.
also set link color so that links are visible.
also add the linkDialog plugin so that it is easier to mess around with links.

it would still be cool if left clicking a link would open it as a child by default.  but with the link dialog you can more easily edit things without accidentally clicking on links.