feat: selectable home page and delete from home
problem: could not click up and down when searching for tabs.
also could not delete tabs from home screen.  deleting tabs is a pain

solution: add selector that can move up and down when searching for tabs
allow delting from the home page.

also added in a lock object on the tabs type. but this does not do
anyhting yet